meta standalone conditional compilation
