+++ framework/meta/Cargo.toml
@@ -19,32 +19,37 @@ categories = ["cryptography::cryptocurrencies", "development-tools::debugging"]
 [[bin]]
 name = "sc-meta"
 path = "src/main.rs"
+required-features = ["standalone"]
 
 [features]
+standalone = ["ruplacer", "reqwest", "zip", "copy_dir", "pathdiff", "common-path"]
 template-test-current = []
 template-test-released = []
 
 [dependencies]
-reqwest = { version = "0.11.4", features = ["blocking", "json"] }
 serde = { version = "1.0", features = ["derive"] }
-tempfile = "3.3.0"
-copy_dir = "0.1.2"
-zip = { version = "0.6.4", features = ["deflate"], default-features = false }
 serde_json = "1.0"
 rustc_version = "0.4"
 toml = { version = "0.5.8", features = ["preserve_order"] }
-ruplacer = { version = "0.8.1", default-features = false }
 clap = { version = "4.1.0", features = ["derive"] }
 colored = "2.0"
-pathdiff = "0.2.1"
-common-path = "1.0.0"
 lazy_static = "1.4.0"
 convert_case = "0.6.0"
 hex = "0.4"
 wasmparser = "0.116.0"
 wasmprinter = "0.2.71"
 
+ruplacer = { version = "0.8.1", default-features = false, optional = true }
+reqwest = { version = "0.11.4", features = ["blocking", "json"], optional = true }
+zip = { version = "0.6.4", features = ["deflate"], default-features = false, optional = true }
+copy_dir = { version = "0.1.2", optional = true }
+pathdiff = { version = "0.2.1", optional = true }
+common-path = { version = "1.0.0", optional = true }
+
 [dependencies.multiversx-sc]
 version = "=0.44.0"
 path = "../base"
 features = ["alloc", "num-bigint", "promises"]
+
+[dev-dependencies]
+multiversx-sc-meta = { path = ".", features = ["standalone"] }
