+++ framework/meta/src/cmd.rs
@@ -1,2 +1,4 @@
 pub mod contract;
+
+#[cfg(feature = "standalone")]
 pub mod standalone;
