+++ framework/meta/src/cmd/standalone/template.rs
@@ -1,5 +1,6 @@
 mod contract_creator;
 mod contract_creator_target;
+mod copy_util;
 mod repo_source;
 mod repo_temp_download;
 mod repo_version;
