+++ framework/meta/src/cmd/standalone/template/contract_creator.rs
