+++ framework/meta/src/cmd/standalone/template/contract_creator_target.rs
