+++ framework/meta/src/cmd/standalone/template/repo_version.rs
