+++ framework/meta/src/cmd/standalone/template/template_metadata.rs
