+++ framework/meta/src/folder_structure.rs
@@ -1,9 +1,7 @@
-mod copy_util;
 mod pretty_print;
 mod relevant_directory;
 mod version_req;
 
-pub use copy_util::*;
 pub use pretty_print::*;
 pub use relevant_directory::*;
 pub use version_req::*;
