+++ framework/meta/src/main.rs
@@ -1,3 +1,3 @@
 fn main() {
-    multiversx_sc_meta::cli_main_standalone();
+    multiversx_sc_meta::cmd::standalone::cli_main_standalone();
 }
